more unnecessary code cleanup? sign me up

+++ BleachHack-Fabric-1.14/src/main/java/bleach/hack/module/mods/Fullbright.java
@@ -1,17 +1,17 @@
 /*
  * This file is part of the BleachHack distribution (https://github.com/BleachDrinker420/bleachhack-1.14/).
  * Copyright (c) 2019 Bleach.
- * 
+ *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
@@ -29,7 +29,7 @@ import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
 
 public class Fullbright extends Module {
-			
+
 	public Fullbright() {
 		super("Fullbright", GLFW.GLFW_KEY_C, Category.RENDER, "Turns your gamma setting up.",
 				new SettingMode("Mode: ", "Gamma", "Potion", "Table"));
@@ -49,15 +49,15 @@ public class Fullbright extends Module {
 				g -= 2;
 			}
 		}
-		
+
 		mc.player.removePotionEffect(StatusEffects.NIGHT_VISION);
 		//Vanilla code to remap light level table.
 		for (int i = 0; i <= 15; ++i) {
-			float float_2 = 1.0F - (float)i / 15.0F;
+			float float_2 = 1.0F - i / 15.0F;
 			mc.world.dimension.getLightLevelToBrightness()[i] = (1.0F - float_2) / (float_2 * 3.0F + 1.0F) * 1.0F + 0.0F;
 		}
 	}
-	
+
 	public void onEnable() {
 		super.onEnable();
